+++ scvi/models/autozivae.py
@@ -2,11 +2,10 @@ import torch
 import torch.nn.functional as F
 from torch.distributions import Normal, Beta, Gamma, kl_divergence as kl
 import numpy as np
-
-
-from scvi.models.log_likelihood import log_zinb_positive, log_nb_positive
-from scvi.models.vae import VAE
 from scipy.special import logit
+
+from scvi.models.distributions import ZeroInflatedNegativeBinomial, NegativeBinomial
+from scvi.models.vae import VAE
 from scvi.models.utils import one_hot
 
 from typing import Dict, Optional, Tuple, Union
@@ -34,9 +33,9 @@ class AutoZIVAE(VAE):
                             When set to ``None'', will be set to 1 - beta_prior if beta_prior is not ``None'',
                             otherwise the prior Beta distribution will be learned on an Empirical Bayes fashion.
         :param beta_prior: Float denoting the beta parameter of the prior Beta distribution of
-                            the zero-inflation Bernoulli parameter. Should be between 0 and 1, not included.
-                            When set to ``None'', will be set to 1 - alpha_prior if alpha_prior is not ``None'',
-                            otherwise the prior Beta distribution will be learned on an Empirical Bayes fashion.
+                           the zero-inflation Bernoulli parameter. Should be between 0 and 1, not included.
+                           When set to ``None'', will be set to 1 - alpha_prior if alpha_prior is not ``None'',
+                           otherwise the prior Beta distribution will be learned on an Empirical Bayes fashion.
         :param minimal_dropout: Float denoting the lower bound of the cell-gene ZI rate in the ZINB component.
                                 Must be non-negative. Can be set to 0 but not recommended as this may make
                                 the mixture problem ill-defined.
@@ -73,7 +72,7 @@ class AutoZIVAE(VAE):
             alpha_prior = 1.0 - beta_prior
 
         # Create parameters for Bernoulli Beta prior and posterior distributions
-        # Each paramer, whose values are in (0,1), is encoded as its logit, in the set of real numbers
+        # Each parameter, whose values are in (0,1), is encoded as its logit, in the set of real numbers
 
         if self.zero_inflation == "gene":
             self.alpha_posterior_logit = torch.nn.Parameter(torch.randn(n_input))
@@ -322,12 +321,16 @@ class AutoZIVAE(VAE):
     ) -> torch.Tensor:
 
         # LLs for NB and ZINB
-        ll_zinb = torch.log(1.0 - bernoulli_params + eps_log) + log_zinb_positive(
-            x, px_rate, px_r, px_dropout
+        ll_zinb = torch.log(
+            1.0 - bernoulli_params + eps_log
+        ) + ZeroInflatedNegativeBinomial(
+            mu=px_rate, theta=px_r, zi_logits=px_dropout
+        ).log_prob(
+            x
         )
-        ll_nb = torch.log(bernoulli_params + eps_log) + log_nb_positive(
-            x, px_rate, px_r
-        )
+        ll_nb = torch.log(bernoulli_params + eps_log) + NegativeBinomial(
+            mu=px_rate, theta=px_r
+        ).log_prob(x)
 
         # Reconstruction loss using a logsumexp-type computation
         ll_max = torch.max(ll_zinb, ll_nb)