+++ scvi/models/vae.py
@@ -5,9 +5,13 @@ import numpy as np
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from torch.distributions import Normal, Poisson, kl_divergence as kl
+from torch.distributions import Normal, kl_divergence as kl
 
-from scvi.models.log_likelihood import log_zinb_positive, log_nb_positive
+from scvi.models.distributions import (
+    ZeroInflatedNegativeBinomial,
+    NegativeBinomial,
+    Poisson,
+)
 from scvi.models.modules import Encoder, DecoderSCVI, LinearDecoderSCVI
 from scvi.models.utils import one_hot
 
@@ -203,9 +207,17 @@ class VAE(nn.Module):
         """
         # Reconstruction Loss
         if self.reconstruction_loss == "zinb":
-            reconst_loss = -log_zinb_positive(x, px_rate, px_r, px_dropout).sum(dim=-1)
+            reconst_loss = (
+                -ZeroInflatedNegativeBinomial(
+                    mu=px_rate, theta=px_r, zi_logits=px_dropout
+                )
+                .log_prob(x)
+                .sum(dim=-1)
+            )
         elif self.reconstruction_loss == "nb":
-            reconst_loss = -log_nb_positive(x, px_rate, px_r).sum(dim=-1)
+            reconst_loss = (
+                -NegativeBinomial(mu=px_rate, theta=px_r).log_prob(x).sum(dim=-1)
+            )
         elif self.reconstruction_loss == "poisson":
             reconst_loss = -Poisson(px_rate).log_prob(x).sum(dim=-1)
         return reconst_loss